Docs + fixes + rename
