+++ test/util.js
@@ -1,16 +1,13 @@
 const freeton = require('../src');
 const logger = require('mocha-logger');
-const { CRYSTAL_AMOUNT } = require('../config/general/constants');
 
 const RootContract = require('../contractWrappers/tip3/rootContract');
-const Wallet = require('../contractWrappers/tip3/walletContract');
-const WalletDeployer = require('../contractWrappers/tip3/walletDeployer');
 const { sleep } = require('../src/utils');
 
 /**
  * String to hex
  * @param {String} str 
- * @returns 
+ * @returns {String}
  */
 function toHex(str) {
     return Buffer.from(str, 'utf8').toString('hex');
@@ -19,6 +16,7 @@ function toHex(str) {
 /**
  * Copy JSON
  * @param {JSON} json 
+ * @returns {Object}
  */
 function copyJSON(json) {
     return JSON.parse(JSON.stringify(json));
@@ -42,7 +40,8 @@ async function sendGrams(giver, address, amount) {
 /**
  * Initial token config creation
  * @param {freeton.TonWrapper} tonInstance 
- * @param {JSON} config 
+ * @param {import('../config/contracts/rootContractParameters').RootTIP3Config} config 
+ * @returns {import('../config/contracts/rootContractParameters').RootTIP3Config}
  */
 function initialTokenSetup(tonInstance, config) {
     let tokenConfig = copyJSON(config);
@@ -61,81 +60,6 @@ function initialTokenSetup(tonInstance, config) {
     return tokenConfig;
 }
 
-/**
- * Initial swap config
- * @param {freeton.TonWrapper} tonInstance 
- * @param {JSON} config 
- * @param {Array} tokens
- */
-function initialSwapSetup(tonInstance, config, tokens) {
-    config.root.keyPair = tonInstance.keys[0];
-    config.root.initParams.ownerPubkey = '0x' + tonInstance.keys[0].public;
-
-    config.pair.keyPair = tonInstance.keys[0];
-    config.pair.initParams.token1 = tokens[0].root.rootContract.address;
-    config.pair.initParams.token2 = tokens[1].root.rootContract.address;
-
-    return config;
-}
-
-/**
- * Deploy TIP-3 token root contract and wallets
- * @param {freeton.TonWrapper} tonInstance 
- * @param {JSON} tokenConfig 
- * @param {freeton.ContractWrapper} giverSC
- */
-async function deployTIP3(tonInstance, tokenConfig, giverSC) {
-    let rootSC;
-    let proxyContract;
-    let wallets = [];
-
-    logger.log('#####################################');
-    logger.log('Initial stage');
-
-    for (let contractId = 0; contractId < tokenConfig.walletsAmount; contractId++) {
-        let walletConfig = tokenConfig.walletsConfig[contractId];
-        wallets.push(new Wallet(tonInstance, walletConfig.config, walletConfig.keys));
-        await wallets[contractId].loadContract();
-    }
-
-    tokenConfig.root.config.initParams.wallet_code = wallets[0].walletContract.code;
-    rootSC = new RootContract(tonInstance, tokenConfig.root.config, tokenConfig.root.keys);
-    await rootSC.loadContract();
-
-    logger.log('Deploying root contract');
-    await rootSC.deployContract();
-
-    logger.log('Loading and deploying proxy contract');
-    proxyContract = new WalletDeployer(tonInstance, {
-        initParams: {},
-        constructorParams: {}
-    }, tokenConfig.root.keys);
-    await proxyContract.loadContract();
-    await proxyContract.deployContract(rootSC.rootContract.address);
-    logger.log(`Proxy contract address: ${proxyContract.walletDeployContract.address}`);
-
-    logger.log('Deploying wallet contracts and sending them tons');
-    for (let contractId = 0; contractId < wallets.length; contractId++) {
-        let walletConfig = wallets[contractId].initParams;
-        await proxyContract.deployWallet(walletConfig.wallet_public_key, walletConfig.owner_address);
-
-        let calculatedAddress = await rootSC.calculateFutureWalletAddress(walletConfig.wallet_public_key, walletConfig.owner_address);
-        wallets[contractId].walletContract.address = calculatedAddress;
-
-        await sendGrams(giverSC, calculatedAddress, CRYSTAL_AMOUNT);
-    }
-
-    logger.log('Minting tokens to wallets');
-    for (let contractId = 0; contractId < wallets.length; contractId++) {
-        await rootSC.mintTokensToWallet(wallets[contractId], (tokenConfig.tokensAmount).toLocaleString('en').replace(/,/g, ''));
-    }
-
-    return {
-        wallets: wallets,
-        root: rootSC
-    }
-}
-
 // _______   ________  __       __   ______   _______   __    __ 
 // /       \ /        |/  |  _  /  | /      \ /       \ /  |  /  |
 // $$$$$$$  |$$$$$$$$/ $$ | / \ $$ |/$$$$$$  |$$$$$$$  |$$ | /$$/ 
@@ -147,32 +71,9 @@ async function deployTIP3(tonInstance, tokenConfig, giverSC) {
 // $$/   $$/ $$$$$$$$/ $$/      $$/  $$$$$$/  $$/   $$/ $$/   $$/ 
 
 /**
- * @typedef SwapPairConfig
- * @type {Object}
- * 
- * @property {}
- */
-
-/**
- * @typedef RootSwapPairConfig
- */
-
-/**
- * @typedef RootSwapPairConfigInitParams
- * @type {Object}
- */
-
-/**
- * @typedef RootSwapPairConfigContructParams
- * @type {Object}
- * 
- * @property {}
- */
-
-/**
  * Deploy TIP-3 token root contract and wallets
  * @param {freeton.TonWrapper} tonInstance 
- * @param {JSON} rootConfig
+ * @param {import('../config/contracts/rootContractParameters').RootTIP3Config} rootConfig
  */
 async function deployTIP3Root(tonInstance, rootConfig) {
     let rootSC;
@@ -191,10 +92,10 @@ async function deployTIP3Root(tonInstance, rootConfig) {
 
 /**
  * create config for root swap pair contract
- * @param {JSON} config 
+ * @param {import('../config/contracts/swapPairContractsConfig').SwapPairConfig} config 
  * @param {String} tip3DeployerAddress
  * @param {freeton.TonWrapper} tonInstance 
- * @returns 
+ * @returns {Promise<import('../config/contracts/swapPairContractsConfig').SwapPairConfig>}
  */
 async function createRootSwapPairConfig(config, tip3DeployerAddress, tonInstance) {
     config.root.keyPair = tonInstance.keys[0];
@@ -207,7 +108,8 @@ async function createRootSwapPairConfig(config, tip3DeployerAddress, tonInstance
 /**
  * create config for tip3 root contract
  * @param {freeton.TonWrapper} tonInstance 
- * @param {JSON} config 
+ * @param {import('../config/contracts/rootContractParameters').RootTIP3Config} config 
+ * @returns {Promise<import('../config/contracts/rootContractParameters').RootTIP3Config>}
  */
 async function initialTokenSetup(tonInstance, config) {
     let rootConfig = copyJSON(config);
@@ -224,7 +126,7 @@ async function initialTokenSetup(tonInstance, config) {
 /**
  * Get codes of tip3 contracts
  * @param {freeton.TonWrapper} tonInstance
- * @returns {JSON} output with Root and wallet codes
+ * @returns {{wallet: String, root: String}} output with Root and wallet codes
  */
 async function getTIP3Codes(tonInstance) {
     let wallet = await freeton.requireContract(tonInstance, 'TONTokenWallet');
